Update peripherals.mdx
After spending a whole weekend tinkering trying to make it work after the how to I capitulated and started measuring.  I dont know why and how it worked out for the person writing this how to but even so I used the exact same parts I was not able to make it work, simply by the fact that there is not enough voltage left after the transistor resulting in the GPS module staying off because of insufficient power.